pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-ai-basic"
version = "0.1.0"
description = "Agentic AI practice"
authors = [{ name = "Vishwas Babar" }]
requires-python = ">=3.9"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.47.0",
    "google-generativeai>=0.8.5",
    "gradio>=3.36.1",
    "ipython>=8.18.1",
    "openai>=2.9.0",
    "pypdf2>=3.0.1",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
